+++ Source/Internal/Graphics/UIColor+Hex.swift
@@ -81,3 +81,4 @@ extension UIColor {
         return String(format: "#%08x", rgb)
     }
 }
+